Fixed Error in find_positive_observations in celeba_helper.py
During sampling pos_examples, the original code uses np.random.choice which by default samples with replacement. This needs to be changed to replace=False so we don't retrieve the same positive sample more than once. Additionally, if sample=True, we must check that the len(pos_examples) > num_examples otherwise we would be trying to take a larger sample than the population

This change was overridden during merging the random triplet loss